Update to support gh-pages
This PR contains some minor updates to permit the use of
gh-pages to render the RFCs. This provides better
rendering of the RFCs (we can alter this in future to
further improve).  Free text-search is also now available

Changes are

- a defined theme in the config file
- minmal set of pages to define a hierachy
- Only effect to RFCs is the requirement to add minor
  front matter

Note: github repo admins will need to enable the gh-page
feature on repo settings. The source should be set to 'master branch'
with no theme chosen.

Signed-off-by: Matthew B White <whitemat@uk.ibm.com>