Clarify RFC final comment period does not require strict majority
Clarify RFC final comment period guidance. Specifically, indicate
that a group of at least three interested maintainers should agree
that an RFC is ready for final comment period. The requirement
for a "majority of maintainers" is an unnecessary burden, since the
final comment period itself allows for objections to be raised. Furthermore
it is unclear what majority means when an RFC impacts multiple
Fabric repositories with different sets of maintainers.

Signed-off-by: David Enyeart <enyeart@us.ibm.com>