Fabric Private Chaincode
This PR proposes the *FPC 1.0* architecture and supersedes the earlier [PR#21](https://github.com/hyperledger/fabric-rfcs/pull/21).
PR#21 proposed intrusive and complex changes for an initial realization of the concept of a (stronlgy) private chaincode.
FPC 1.0 takes a different approach as it covers a smaller application domain but thereby avoids any change to the core Fabric code. How to broaden to the application domain covered in PR#21 is outlined in a roadmap of future work.

Signed-off-by: Marcus Brandenburger <bur@zurich.ibm.com>
Co-authored-by: Michael Steiner <michael.steiner@intel.com>
Co-authored-by: Bruno Vavala <bruno.vavala@intel.com>
Co-authored-by: Jeb Linton <jrlinton@us.ibm.com>