BFT signatures: update
The RFC currently calls for adding a gRPC service for getting the Header+Signature stream (Block attestation).

After reviewing the current (SmartBFT fabric v2.3) and proposed solution, it is clear that adding a gRPC service
and multiple new proto messages needlessly duplicates code and bloats the code base. The task is realised in the
current implementation by 6 lines of code - 3 in the orderer (set nil to the block payload, depending on the
content type requested) and 3 in the peer (set the required content type).

We therefore revert this part of the proposal and will simply migrate the code from SmartBFT fabric v2.3 for this functionality.

Signed-off-by: Yoav Tock <tock@il.ibm.com>